requirements.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0

# Validación
pydantic==2.5.0

# Autenticación y seguridad
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilidades
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==40.0.2

# Desarrollo (opcional)
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
